refactor(common): Rename `store_locks` to `cross_process_lock`.
