+++ crates/matrix-sdk-base/src/event_cache/store/memory_store.rs
@@ -19,11 +19,11 @@ use std::{
 
 use async_trait::async_trait;
 use matrix_sdk_common::{
+    cross_process_lock::memory_store_helper::try_take_leased_lock,
     linked_chunk::{
         ChunkIdentifier, ChunkIdentifierGenerator, ChunkMetadata, LinkedChunkId, Position,
         RawChunk, Update, relational::RelationalLinkedChunk,
     },
-    store_locks::memory_store_helper::try_take_leased_lock,
 };
 use ruma::{EventId, OwnedEventId, RoomId, events::relation::RelationType, time::Instant};
 use tracing::error;
