+++ crates/matrix-sdk-base/src/media/store/memory_store.rs
@@ -20,7 +20,7 @@ use std::{
 
 use async_trait::async_trait;
 use matrix_sdk_common::{
-    ring_buffer::RingBuffer, store_locks::memory_store_helper::try_take_leased_lock,
+    cross_process_lock::memory_store_helper::try_take_leased_lock, ring_buffer::RingBuffer,
 };
 use ruma::{
     MxcUri, OwnedMxcUri,
