+++ crates/matrix-sdk/src/encryption/mod.rs
@@ -87,8 +87,8 @@ use self::{
 use crate::{
     attachment::Thumbnail,
     client::{ClientInner, WeakClient},
+    cross_process_lock::CrossProcessLockGuard,
     error::HttpResult,
-    store_locks::CrossProcessLockGuard,
     Client, Error, HttpError, Result, Room, RumaApiError, TransmissionProgress,
 };
